Split up user and post tests per review
